Upgrade turf libs to 7.3.1 (#50)
This appears to fix an intermittent issue with route shape clipping not
working with a "Coordinates must contain numbers" error. The upgrade
fixes a particular case of this route shape bug I've been tracking, and
our issue sounds very similar to this Turf bug:

  https://github.com/Turfjs/turf/issues/2939

for which a fix was shipped in Turf 7.3.0.